Added compatibility for linux operating systems, and a couple of minor changes.
The functions kbhit(), sleep(seconds) were replaced by their linux
counterparts.

Also removed a minor bug that prevented decreasing the speed of the
snake. This was due to repititive usage of 'd'. (fixed using 'g')

Another bug was that "r+" mode was used in file handling, replaced by
"a+".
This is beacuse r+ returns NULL if the file does not exist.

Known bugs:
Arrow key input not supported in linux.